requirements.txt
Flask==2.0.1
torchvision==0.10.0
torch==1.9.0
numpy==1.21.1
opencv-python==4.5.3.56
albumentations==1.0.3
Pillow==8.3.1
timm==0.6.11